# api/openapi.yaml
openapi: 3.1.0
info:
  title: Araru Server API
  version: 1.0.0
  description: Stable Araru API v1. Product endpoints are exclusively exposed under /api/v1.
servers:
  - url: http://localhost:3001/api
    description: Araru API base; product paths are versioned under /v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: araru_session
...
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
paths:
...
  /v1/admin/storage/providers/{id}:
    get:
      operationId: getStorageProvider
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { description: Provider capabilities } }
    patch:
      operationId: patchStorageProvider
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { description: Provider settings saved } }
    delete:
      operationId: deleteStorageProvider
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '204': { description: Settings deleted } }
...
  /v1/admin/storage/r2/upload-url:
    post:
      operationId: createR2UploadUrl
      security: [{ sessionCookie: [] }]
      responses: { '201': { description: Temporary signed upload URL } }
  /v1/admin/storage/r2/complete:
    post:
      operationId: completeR2Upload
      security: [{ sessionCookie: [] }]
      responses: { '201': { description: Registered R2 object } }